"started a mini digital banking app, written some business logic in the account class."

+++ src/main/java/com/Joysbright/digiBank/model/Account.java
@@ -13,7 +13,7 @@ import java.util.List;
 @Builder
 public class Account {
 
-    public int minimumBalance;
+    public double minimumBalance;
 
     @Generated()
     private String accountId;
